hotel-management-booking/requirements.txt
fastapi
uvicorn
mysql-connector-python
requests
python-dotenv
